fix: Seperate roomescape and reservation test
